Small clean for redundant examples
Previous examples are now redundant with the inclusion of classes. These have been moved into a seperate folder.

Note: a single point in the DCSM tide test case fails to interpolate properly. This must be fixed. No leads on potential causes at the moment, but it is close to shore. Otherwise the test passes for Guayaquil and DCSM test case.

README must also be updated.